refactor: update import paths in test files for improved readability
