+++ test/unit/data/configuration/impl/local-config-source.test.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {expect} from 'chai';
-import {LocalConfigSource} from '../../../../../../src/data/configuration/impl/local-config-source.js';
+import {LocalConfigSource} from '../../../../../src/data/configuration/impl/local-config-source.js';
 
 describe('LocalConfigSource', () => {
   it('constructor with null ObjectMapper throws error', () => {
